Extract Subscription methods to their very own Saasramp::Subscription module.
This allows for easier customization in the application by letting users create their own Subscription model and include Saasramp::Subscription into it.

That in turn lets them define their own methods on it and add observers to the model without running into "A copy of SubscriptionObserver has been removed from the tree, but is still active" errors.