SubscriptionPlan#prorated_value should never return an amount larger than the plan rate.
Higher values than the rate can occur if next_renewal_on has been moved manually or if there are 31 days left on a subscription.

The first case means the subscriber hasn't actually paid for those extra days anyways, so they shouldn't be prorated.

The last case would allow a subscriber to switch between a free and a paid plan in a month with 31 days and get an extra 1/30 of the price refunded on every switch back.